requirements.txt
# requirements.txt for Silica-X (Theme: Ember)
aiohttp
cryptography>=42.0.0
matplotlib>=3.8.0
Pillow>=10.0.0
python-docx>=1.1.0
pytesseract>=0.3.10
reportlab>=4.2.0
scapy>=2.5.0
rich # optional
